@@ -16,12 +16,11 @@ classifiers = [
     "Topic :: Scientific/Engineering :: Bio-Informatics",
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
 ]
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 description = "VICC normalization routines for genes"
 license = {file = "LICENSE"}
 dependencies = [
@@ -30,7 +29,7 @@ dependencies = [
     "uvicorn",
     "click",
     "boto3",
-    "ga4gh.vrs==2.0.0a3",
+    "ga4gh.vrs~=2.0.0a8",
 ]
 dynamic = ["version"]
 
@@ -38,7 +37,7 @@ dynamic = ["version"]
 pg = ["psycopg[binary]"]
 etl = ["gffutils", "biocommons.seqrepo", "wags-tails>=0.1.1"]
 test = ["pytest>=6.0", "pytest-cov", "mock", "httpx"]
-dev = ["pre-commit", "ruff==0.2.0"]
+dev = ["pre-commit", "ruff==0.5.0"]
 docs = [
     "sphinx==6.1.3",
     "sphinx-autodoc-typehints==1.22.0",
@@ -108,16 +107,22 @@ select = [
     "DTZ",  # https://docs.astral.sh/ruff/rules/#flake8-datetimez-dtz
     "T10",  # https://docs.astral.sh/ruff/rules/#flake8-datetimez-dtz
     "EM",  # https://docs.astral.sh/ruff/rules/#flake8-errmsg-em
+    "LOG",  # https://docs.astral.sh/ruff/rules/#flake8-logging-log
     "G",  # https://docs.astral.sh/ruff/rules/#flake8-logging-format-g
+    "INP",  # https://docs.astral.sh/ruff/rules/#flake8-no-pep420-inp
     "PIE",  # https://docs.astral.sh/ruff/rules/#flake8-pie-pie
     "T20",  # https://docs.astral.sh/ruff/rules/#flake8-print-t20
     "PT",  # https://docs.astral.sh/ruff/rules/#flake8-pytest-style-pt
     "Q",  # https://docs.astral.sh/ruff/rules/#flake8-quotes-q
     "RSE",  # https://docs.astral.sh/ruff/rules/#flake8-raise-rse
     "RET",  # https://docs.astral.sh/ruff/rules/#flake8-return-ret
+    "SLF",  # https://docs.astral.sh/ruff/rules/#flake8-self-slf
     "SIM",  # https://docs.astral.sh/ruff/rules/#flake8-simplify-sim
+    "ARG",  # https://docs.astral.sh/ruff/rules/#flake8-unused-arguments-arg
     "PTH",  # https://docs.astral.sh/ruff/rules/#flake8-use-pathlib-pth
     "PGH",  # https://docs.astral.sh/ruff/rules/#pygrep-hooks-pgh
+    "PERF",  # https://docs.astral.sh/ruff/rules/#perflint-perf
+    "FURB",  # https://docs.astral.sh/ruff/rules/#refurb-furb
     "RUF",  # https://docs.astral.sh/ruff/rules/#ruff-specific-rules-ruf
 ]
 fixable = [
@@ -128,15 +133,19 @@ fixable = [
     "ANN",
     "B",
     "C4",
+    "LOG",
     "G",
     "PIE",
     "PT",
     "RSE",
     "SIM",
+    "PERF",
+    "FURB",
     "RUF"
 ]
-# ANN101 - missing-type-self
 # ANN003 - missing-type-kwargs
+# ANN101 - missing-type-self
+# ANN102 - missing-type-cls
 # D203 - one-blank-line-before-class
 # D205 - blank-line-after-summary
 # D206 - indent-with-spaces*
@@ -152,7 +161,7 @@ fixable = [
 # S321 - suspicious-ftp-lib-usage
 # *ignored for compatibility with formatter
 ignore = [
-    "ANN101", "ANN003",
+    "ANN003", "ANN101", "ANN102",
     "D203", "D205", "D206", "D213", "D300", "D400", "D415",
     "E111", "E114", "E117", "E501",
     "W191",
@@ -170,7 +179,10 @@ ignore = [
 # D100 - undocumented-public-module
 # D103 - undocumented-public-function
 # I001 - unsorted-imports
-"tests/*" = ["ANN001", "ANN2", "ANN102", "S101", "B011"]
+# INP001 - implicit-namespace-package
+# ARG001 - unused-function-argument
+# SLF001 - private-member-acces
+"tests/*" = ["ANN001", "ANN2", "ANN102", "S101", "INP001", "SLF001", "ARG001"]
 "*__init__.py" = ["F401"]
 "gene/schemas.py" = ["ANN001", "ANN201", "N805"]
 "docs/source/conf.py" = ["D100", "I001", "D103", "ANN201", "ANN001"]
