@@ -1,10 +1,12 @@
 """Provide DynamoDB client."""
+
 import atexit
 import logging
 import sys
+from collections.abc import Generator
 from os import environ
 from pathlib import Path
-from typing import Any, Dict, Generator, List, Optional, Set, Union
+from typing import Any
 
 import boto3
 import click
@@ -25,13 +27,13 @@ from gene.database.database import (
 )
 from gene.schemas import RecordType, RefType, SourceMeta, SourceName
 
-logger = logging.getLogger(__name__)
+_logger = logging.getLogger(__name__)
 
 
 class DynamoDbDatabase(AbstractDatabase):
     """Database class employing DynamoDB."""
 
-    def __init__(self, db_url: Optional[str] = None, **db_args) -> None:
+    def __init__(self, db_url: str | None = None, **db_args) -> None:
         """Initialize Database class.
 
         :param str db_url: URL endpoint for DynamoDB source
@@ -87,7 +89,7 @@ class DynamoDbDatabase(AbstractDatabase):
         self._cached_sources = {}
         atexit.register(self.close_connection)
 
-    def list_tables(self) -> List[str]:
+    def list_tables(self) -> list[str]:
         """Return names of tables in database.
 
         :return: Table names in DynamoDB
@@ -154,7 +156,7 @@ class DynamoDbDatabase(AbstractDatabase):
         existing_tables = self.list_tables()
         exists = self.gene_table in existing_tables
         if not exists:
-            logger.info("%s table is missing or unavailable.", self.gene_table)
+            _logger.info("%s table is missing or unavailable.", self.gene_table)
         return exists
 
     def check_tables_populated(self) -> bool:
@@ -171,7 +173,7 @@ class DynamoDbDatabase(AbstractDatabase):
             KeyConditionExpression=Key("item_type").eq("source"),
         ).get("Items", [])
         if len(sources) < len(SourceName):
-            logger.info("Gene sources table is missing expected sources.")
+            _logger.info("Gene sources table is missing expected sources.")
             return False
 
         records = self.genes.query(
@@ -180,7 +182,7 @@ class DynamoDbDatabase(AbstractDatabase):
             Limit=1,
         )
         if len(records.get("Items", [])) < 1:
-            logger.info("Gene records index is empty.")
+            _logger.info("Gene records index is empty.")
             return False
 
         normalized_records = self.genes.query(
@@ -189,7 +191,7 @@ class DynamoDbDatabase(AbstractDatabase):
             Limit=1,
         )
         if len(normalized_records.get("Items", [])) < 1:
-            logger.info("Normalized gene records index is empty.")
+            _logger.info("Normalized gene records index is empty.")
             return False
 
         return True
@@ -199,7 +201,7 @@ class DynamoDbDatabase(AbstractDatabase):
         if not self.check_schema_initialized():
             self._create_genes_table()
 
-    def get_source_metadata(self, src_name: Union[str, SourceName]) -> Dict:
+    def get_source_metadata(self, src_name: str | SourceName) -> dict:
         """Get license, versioning, data lookup, etc information for a source.
 
         :param src_name: name of the source to get data for
@@ -222,7 +224,7 @@ class DynamoDbDatabase(AbstractDatabase):
 
     def get_record_by_id(
         self, concept_id: str, case_sensitive: bool = True, merge: bool = False
-    ) -> Optional[Dict]:
+    ) -> dict | None:
         """Fetch record corresponding to provided concept ID
 
         :param str concept_id: concept ID for gene record
@@ -250,7 +252,7 @@ class DynamoDbDatabase(AbstractDatabase):
             del record["label_and_type"]
             return record
         except ClientError as e:
-            logger.error(
+            _logger.error(
                 "boto3 client error on get_records_by_id for search term %s: %s",
                 concept_id,
                 e.response["Error"]["Message"],
@@ -259,7 +261,7 @@ class DynamoDbDatabase(AbstractDatabase):
         except (KeyError, IndexError):  # record doesn't exist
             return None
 
-    def get_refs_by_type(self, search_term: str, ref_type: RefType) -> List[str]:
+    def get_refs_by_type(self, search_term: str, ref_type: RefType) -> list[str]:
         """Retrieve concept IDs for records matching the user's query. Other methods
         are responsible for actually retrieving full records.
 
@@ -273,14 +275,14 @@ class DynamoDbDatabase(AbstractDatabase):
             matches = self.genes.query(KeyConditionExpression=filter_exp)
             return [m["concept_id"] for m in matches.get("Items", None)]
         except ClientError as e:
-            logger.error(
+            _logger.error(
                 "boto3 client error on get_refs_by_type for search term %s: %s",
                 search_term,
                 e.response["Error"]["Message"],
             )
             return []
 
-    def get_all_concept_ids(self) -> Set[str]:
+    def get_all_concept_ids(self) -> set[str]:
         """Retrieve concept IDs for use in generating normalized records.
 
         :return: List of concept IDs as strings.
@@ -298,14 +300,13 @@ class DynamoDbDatabase(AbstractDatabase):
             else:
                 response = self.genes.scan(**params)
             records = response["Items"]
-            for record in records:
-                concept_ids.append(record["concept_id"])
+            concept_ids.extend(record["concept_id"] for record in records)
             last_evaluated_key = response.get("LastEvaluatedKey")
             if not last_evaluated_key:
                 break
         return set(concept_ids)
 
-    def get_all_records(self, record_type: RecordType) -> Generator[Dict, None, None]:
+    def get_all_records(self, record_type: RecordType) -> Generator[dict, None, None]:
         """Retrieve all source or normalized records. Either return all source records,
         or all records that qualify as "normalized" (i.e., merged groups + source
         records that are otherwise ungrouped).
@@ -363,7 +364,7 @@ class DynamoDbDatabase(AbstractDatabase):
         except ClientError as e:
             raise DatabaseWriteException(e) from e
 
-    def add_record(self, record: Dict, src_name: SourceName) -> None:
+    def add_record(self, record: dict, src_name: SourceName) -> None:
         """Add new record to database.
 
         :param Dict record: record to upload
@@ -377,7 +378,7 @@ class DynamoDbDatabase(AbstractDatabase):
         try:
             self.batch.put_item(Item=record)
         except ClientError as e:
-            logger.error(
+            _logger.error(
                 "boto3 client error on add_record for %s: %s",
                 concept_id,
                 e.response["Error"]["Message"],
@@ -396,7 +397,7 @@ class DynamoDbDatabase(AbstractDatabase):
                         item, record["concept_id"], item_type, src_name
                     )
 
-    def add_merged_record(self, record: Dict) -> None:
+    def add_merged_record(self, record: dict) -> None:
         """Add merged record to database.
 
         :param record: merged record to add
@@ -410,7 +411,7 @@ class DynamoDbDatabase(AbstractDatabase):
         try:
             self.batch.put_item(Item=record)
         except ClientError as e:
-            logger.error(
+            _logger.error(
                 "boto3 client error on add_record for " "%s: %s",
                 concept_id,
                 e.response["Error"]["Message"],
@@ -437,7 +438,7 @@ class DynamoDbDatabase(AbstractDatabase):
         try:
             self.batch.put_item(Item=record)
         except ClientError as e:
-            logger.error(
+            _logger.error(
                 "boto3 client error adding reference %s for %s with match type %s: %s",
                 term,
                 concept_id,
@@ -472,7 +473,7 @@ class DynamoDbDatabase(AbstractDatabase):
                 )
                 raise DatabaseWriteException(err_msg) from e
 
-            logger.error(
+            _logger.error(
                 "boto3 client error in `database.update_record()`: %s",
                 e.response["Error"]["Message"],
             )
@@ -531,16 +532,16 @@ class DynamoDbDatabase(AbstractDatabase):
             with self.genes.batch_writer(
                 overwrite_by_pkeys=["label_and_type", "concept_id"]
             ) as batch:
-                for record in records:
-                    try:
+                try:
+                    for record in records:
                         batch.delete_item(
                             Key={
                                 "label_and_type": record["label_and_type"],
                                 "concept_id": record["concept_id"],
                             }
                         )
-                    except ClientError as e:
-                        raise DatabaseWriteException(e) from e
+                except ClientError as e:
+                    raise DatabaseWriteException(e) from e
 
     def complete_write_transaction(self) -> None:
         """Conclude transaction or batch writing if relevant."""
@@ -551,7 +552,7 @@ class DynamoDbDatabase(AbstractDatabase):
         """Perform any manual connection closure procedures if necessary."""
         self.batch.__exit__(*sys.exc_info())
 
-    def load_from_remote(self, url: Optional[str] = None) -> None:
+    def load_from_remote(self, url: str | None = None) -> None:
         """Load DB from remote dump. Not available for DynamoDB database backend.
 
         :param url: remote location to retrieve gzipped dump file from
