@@ -1,9 +1,9 @@
 """A base class for extraction, transformation, and loading of data."""
+
 import logging
 import re
 from abc import ABC, abstractmethod
 from pathlib import Path
-from typing import Dict, List, Optional, Union
 
 import click
 import pydantic
@@ -15,8 +15,7 @@ from gene import ITEM_TYPES, SEQREPO_ROOT_DIR
 from gene.database import AbstractDatabase
 from gene.schemas import Gene, GeneSequenceLocation, MatchType, SourceName
 
-logger = logging.getLogger("gene")
-logger.setLevel(logging.DEBUG)
+_logger = logging.getLogger(__name__)
 
 
 DATA_DISPATCH = {
@@ -33,7 +32,7 @@ class Base(ABC):
         self,
         database: AbstractDatabase,
         seqrepo_dir: Path = SEQREPO_ROOT_DIR,
-        data_path: Optional[Path] = None,
+        data_path: Path | None = None,
         silent: bool = True,
     ) -> None:
         """Instantiate Base class.
@@ -51,8 +50,8 @@ class Base(ABC):
         self._processed_ids = []
 
     def _get_data_handler(
-        self, data_path: Optional[Path] = None
-    ) -> Union[HgncData, EnsemblData, NcbiGeneData]:
+        self, data_path: Path | None = None
+    ) -> HgncData | EnsemblData | NcbiGeneData:
         """Construct data handler instance for source. Overwrite for edge-case sources.
 
         :param data_path: location of data storage
@@ -60,7 +59,7 @@ class Base(ABC):
         """
         return DATA_DISPATCH[self._src_name](data_dir=data_path, silent=self._silent)
 
-    def perform_etl(self, use_existing: bool = False) -> List[str]:
+    def perform_etl(self, use_existing: bool = False) -> list[str]:
         """Public-facing method to begin ETL procedures on given data.
         Returned concept IDs can be passed to Merge method for computing
         merged concepts.
@@ -70,6 +69,7 @@ class Base(ABC):
             uploaded.
         """
         self._extract_data(use_existing)
+        _logger.info("Transforming and loading %s data to DB...", self._src_name.value)
         if not self._silent:
             click.echo("Transforming and loading data to DB...")
         self._add_meta()
@@ -88,6 +88,7 @@ class Base(ABC):
         self._data_file, self._version = self._data_source.get_latest(
             from_local=use_existing
         )
+        _logger.info("Acquired data for %s: %s", self._src_name.value, self._data_file)
 
     @abstractmethod
     def _transform_data(self) -> None:
@@ -99,7 +100,7 @@ class Base(ABC):
         """Add source meta to database source info."""
         raise NotImplementedError
 
-    def _load_gene(self, gene: Dict) -> None:
+    def _load_gene(self, gene: dict) -> None:
         """Load a gene record into database. This method takes responsibility for:
          * validating structure correctness
          * removing duplicates from list-like fields
@@ -110,7 +111,7 @@ class Base(ABC):
         try:
             Gene(match_type=MatchType.NO_MATCH, **gene)
         except pydantic.ValidationError as e:
-            logger.warning("Unable to load %s due to validation error: %s", gene, e)
+            _logger.warning("Unable to load %s due to validation error: %s", gene, e)
         else:
             concept_id = gene["concept_id"]
             gene["label_and_type"] = f"{concept_id.lower()}##identity"
@@ -140,7 +141,7 @@ class Base(ABC):
             raise NotADirectoryError(err_msg)
         return SeqRepo(seqrepo_dir)
 
-    def _set_cl_interval_range(self, loc: str, arm_ix: int, location: Dict) -> None:
+    def _set_cl_interval_range(self, loc: str, arm_ix: int, location: dict) -> None:
         """Set the Chromosome location interval range.
 
         :param loc: A gene location
@@ -202,7 +203,7 @@ class Base(ABC):
     #             return chr_location
     #     return None
 
-    def _get_seq_id_aliases(self, seq_id: str) -> List[str]:
+    def _get_seq_id_aliases(self, seq_id: str) -> list[str]:
         """Get GA4GH aliases for a sequence id
 
         :param seq_id: Sequence ID accession
@@ -212,10 +213,10 @@ class Base(ABC):
         try:
             aliases = self.seqrepo.translate_alias(seq_id, target_namespaces="ga4gh")
         except KeyError as e:
-            logger.warning("SeqRepo raised KeyError: %s", e)
+            _logger.warning("SeqRepo raised KeyError: %s", e)
         return aliases
 
-    def _get_sequence_location(self, seq_id: str, gene: Feature, params: Dict) -> Dict:
+    def _get_sequence_location(self, seq_id: str, gene: Feature, params: dict) -> dict:
         """Get a gene's GeneSequenceLocation.
 
         :param seq_id: The sequence ID.
@@ -239,7 +240,7 @@ class Base(ABC):
                     sequence_id=sequence,
                 ).model_dump()
             else:
-                logger.warning(
+                _logger.warning(
                     "%s has invalid interval: start=%i end=%i",
                     params["concept_id"],
                     gene.start - 1,
