@@ -1,8 +1,8 @@
 """Defines the HGNC ETL methods."""
+
 import json
 import logging
 import re
-from typing import Dict
 
 from gene import PREFIX_LOOKUP
 from gene.etl.base import Base
@@ -18,8 +18,7 @@ from gene.schemas import (
     SymbolStatus,
 )
 
-logger = logging.getLogger("gene")
-logger.setLevel(logging.DEBUG)
+_logger = logging.getLogger(__name__)
 
 
 class HGNC(Base):
@@ -27,7 +26,7 @@ class HGNC(Base):
 
     def _transform_data(self) -> None:
         """Transform the HGNC source."""
-        logger.info("Transforming HGNC...")
+        _logger.info("Transforming HGNC...")
         with self._data_file.open() as f:
             data = json.load(f)
 
@@ -58,9 +57,9 @@ class HGNC(Base):
             if "locus_type" in r:
                 gene["gene_type"] = r["locus_type"]
                 self._load_gene(gene)
-        logger.info("Successfully transformed HGNC.")
+        _logger.info("Successfully transformed HGNC.")
 
-    def _get_aliases(self, r: Dict, gene: Dict) -> None:
+    def _get_aliases(self, r: dict, gene: dict) -> None:
         """Store aliases in a gene record.
 
         :param r: A gene record in the HGNC data file
@@ -77,7 +76,7 @@ class HGNC(Base):
         if alias_symbol or enzyme_id:
             gene["aliases"] = list(set(alias_symbol + enzyme_id))
 
-    def _get_previous_symbols(self, r: Dict, gene: Dict) -> None:
+    def _get_previous_symbols(self, r: dict, gene: dict) -> None:
         """Store previous symbols in a gene record.
 
         :param r: A gene record in the HGNC data file
@@ -87,7 +86,7 @@ class HGNC(Base):
         if prev_symbols:
             gene["previous_symbols"] = list(set(prev_symbols))
 
-    def _get_xrefs_associated_with(self, r: Dict, gene: Dict) -> None:
+    def _get_xrefs_associated_with(self, r: dict, gene: dict) -> None:
         """Store xrefs and/or associated_with refs in a gene record.
 
         :param r: A gene record in the HGNC data file
@@ -139,7 +138,7 @@ class HGNC(Base):
                     else:
                         self._get_xref_associated_with(key, src, r, associated_with)
                 else:
-                    logger.warning("%s not in schemas.py", key)
+                    _logger.warning("%s not in schemas.py", key)
 
         if xrefs:
             gene["xrefs"] = xrefs
@@ -147,7 +146,7 @@ class HGNC(Base):
             gene["associated_with"] = associated_with
 
     def _get_xref_associated_with(
-        self, key: str, src: str, r: Dict, src_type: Dict
+        self, key: str, src: str, r: dict, src_type: dict
     ) -> None:
         """Add an xref or associated_with ref to a gene record.
 
@@ -164,7 +163,7 @@ class HGNC(Base):
                 r[src] = r[src].split(":")[-1].strip()
             src_type.append(f"{NamespacePrefix[key.upper()].value}" f":{r[src]}")
 
-    def _get_location(self, r: Dict, gene: Dict) -> None:
+    def _get_location(self, r: dict, gene: dict) -> None:
         """Store GA4GH VRS ChromosomeLocation in a gene record.
         https://vr-spec.readthedocs.io/en/1.1/terms_and_model.html#chromosomelocation
 
@@ -198,7 +197,7 @@ class HGNC(Base):
         if not gene["location_annotations"]:
             del gene["location_annotations"]
 
-    def _set_annotation(self, loc: str, gene: Dict) -> None:
+    def _set_annotation(self, loc: str, gene: dict) -> None:
         """Set the annotations attribute if one is provided.
         Return `True` if a location is provided, `False` otherwise.
 
@@ -217,7 +216,7 @@ class HGNC(Base):
                     return None
         return loc
 
-    def _set_location(self, loc: str, location: Dict, gene: Dict) -> None:
+    def _set_location(self, loc: str, location: dict, gene: dict) -> None:
         """Set a gene's location.
 
         :param loc: A gene location
