@@ -1,2 +1,3 @@
 """Gene normalizer version"""
-__version__ = "0.3.0-dev1"
+
+__version__ = "0.3.2"
