@@ -1,4 +1,5 @@
 """Test the emit_warnings function."""
+
 from gene.database import create_db
 from gene.query import QueryHandler
 
@@ -21,7 +22,7 @@ def test_emit_warnings():
     actual_warnings = query_handler._emit_warnings("sp ry3")
     assert actual_warnings == actual_warnings
 
-    actual_warnings = query_handler._emit_warnings("sp\u00A0ry3")
+    actual_warnings = query_handler._emit_warnings("sp\u00a0ry3")
     assert expected_warnings == actual_warnings
 
     actual_warnings = query_handler._emit_warnings("sp&nbsp;ry3")
